+++ CodesHunter.py
@@ -1,8 +1,8 @@
 import os.path
 import Globals as gs
 from time import sleep
-from Tools.fileTools import ControlFile, GetDateModified, GetModifiedFiles
-from Tools.xmlTools import GetAllFiles, ReadControlFile, PopulateCodes2, ControlFileUpdate, CleanControlEntries
+from Tools.fileTools import ControlFile, LoadJson, SaveJson, CleanControlEntries, ControlFileUpdate, GetDateModified, GetModifiedFiles
+from Tools.xmlTools import GetAllFiles, ReadControlFile, PopulateCodes2
 from Tools.errorHandler import ErrorLogging
 
 
@@ -13,19 +13,13 @@ def CodesHunter():
         logFiles = []
         currentDateFiles = {}
         previousDateFiles = {}
-        controlData = []
         logFilesToCheck = []
-        
+
         # Make sure the Control file exists
-        controlFile = ControlFile()
+        gs.controlJson = ControlFile()
         
         # Clean the control file of non existent log files
-        CleanControlEntries(controlFile, GetAllFiles(gs.configList['vrchat-log-path']+'output_log_*.txt'))
-
-        # Initialize with control file's dates
-        controlData = ReadControlFile(controlFile)
-        for item in controlData:
-            currentDateFiles[item[0]] = item[1]                         # 0 File, 1 Date
+        CleanControlEntries(gs.controlJson, GetAllFiles(gs.configList['vrchat-log-path']+'output_log_*.txt'))
 
         while True:
             # Let's retrieve all files again, just in case VRC crashed or was restarted, it creates a new log file
@@ -35,7 +29,7 @@ def CodesHunter():
             currentDateFiles = GetDateModified(logFiles)                # Obtain the new dates
             if gs.forceRefreshCodes:                                    # In case of a Code.xml file corruption this will force a refresh
                 previousDateFiles = []
-                controlData = ReadControlFile(controlFile)              # Refresh the control data to update the cursors
+                gs.controlJson = ControlFile()                             # Refresh the control data to update the cursors
                 gs.forceRefreshCodes = False
             if currentDateFiles != previousDateFiles:
                 print('Modified files detected, checking...')
@@ -45,16 +39,15 @@ def CodesHunter():
                 
                 for log in logFilesToCheckSorted:
                     cursorOld = '0'
-                    for item in controlData:
-                        if item[0]==log[0]:
-                            cursorOld = item[2]
-                            break
+                    if log[0] in gs.controlJson:
+                        cursorOld = gs.controlJson[log[0]]["cursor"]
+
                     cursorAux = PopulateCodes2(log[0], gs._FOLDER_CODES, cursorOld)
-                    ControlFileUpdate(controlFile, log[0], log[1], cursorAux)
-                    controlData = ReadControlFile(controlFile)
+                    ControlFileUpdate(gs.controlJson, log[0], log[1], cursorAux)
+                    SaveJson(gs._FILE_CONTROLJSON, gs.controlJson)
                     print('Updated Control file')
 
             sleep( float(gs.configList['file-delay']) )                      # The program is very eepy, let it rest
     except Exception as e:
         print(e)
-        ErrorLogging(f"Error in CodesHunter2: {e}")
+        ErrorLogging(f"Error in CodesHunter: {e}")