+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 **Changes**
 - Removed `codes-folder` from config.xml.default
 - Cleaned up code and fixed references to the Codes folder.
+- Moved control.xml from xml to json format. [#63](https://github.com/69MichelleDB/ToNCodes/issues/63)
+    - Fixes to prevent null cursors when parsing Log files.
 
 
 ## [alpha-0.7.9] - 2025.04.11